requirements.txt
discord.py
python-dotenv
gspread
oauth2client
pyyaml
psycopg2-binary
requests
tzdata
rapidfuzz
aiohttp
colorama
ruamel.yaml
fastapi
uvicorn[standard]
sqlalchemy
alembic
websockets
redis
jinja2
python-jose[cryptography]